refactor: split everythin into mod
