+++ examples/simple_service.rs
@@ -1,8 +1,7 @@
 use girolle::prelude::*;
-use girolle::{Result, RpcService, RpcTask};
 use serde_json;
 
-fn hello(s: &[Value]) -> Result<Value> {
+fn hello(s: &[Value]) -> NamekoResult<Value> {
     // Parse the incomming data
     let n: String = serde_json::from_value(s[0].clone())?;
     let hello_str: Value = format!("Hello, {}!, by Girolle", n).into();
@@ -26,7 +25,7 @@ fn fibonacci(n: u64) -> u64 {
     }
 }
 
-fn fibonacci_reccursive(s: &[Value]) -> Result<Value> {
+fn fibonacci_reccursive(s: &[Value]) -> NamekoResult<Value> {
     let n: u64 = serde_json::from_value(s[0].clone())?;
     let result: Value = serde_json::to_value(fibonacci(n))?;
     Ok(result)
