+++ girolle/benches/macro.rs
@@ -1,6 +1,7 @@
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 
 use girolle_macro::girolle;
+use girolle::types::NamekoResult;
 use serde_json::Value;
 
 fn fibonacci_fast(n: u64) -> u64 {
